# docs/swagger.yaml
...
openapi: 3.1.0
...
paths:
...
  /api/loan/{id}:
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Create Loan Request
      tags:
      - Loan Resource
  /api/loan/{id}/contact:
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Bad Request
      summary: Notify Owner
      tags:
      - Loan Resource
  /api/loan/{id}/extend:
    patch:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: integer
                format: int32
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Extend Loan
      tags:
      - Loan Resource
  /api/loan/{id}/return:
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Bad Request
      summary: Close Loan
      tags:
      - Loan Resource
  /api/loan/{id}/start:
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Start Loan
      tags:
      - Loan Resource
  /api/loan/{id}/status:
    patch:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Manage Request
      tags:
      - Loan Resource
...
info:
  title: bibliomap-api API
  version: 1.0.0-SNAPSHOT